fix: [NUTR-50] - exported new type and fixes function
