+++ src/types/dataTotalType.ts
@@ -1,0 +1,7 @@
+export interface DataTotalType {
+  name: string;
+  carbohydrate: number;
+  protein: number;
+  lipid: number;
+  energy: number;
+}
